API Documentation

Docstrings for Dojo.jl interface members can be accessed through Julia's built-in documentation system or in the list below.

Contents

API Documentation

Contents

Index

Mechanism

Representation

Environments

Simulation

Gradients

Visualization

Index

Dojo.Body

Dojo.Contact

Dojo.ContactConstraint

Dojo.ImpactContact

Dojo.JointConstraint

Dojo.LinearContact

Dojo.NonlinearContact

Dojo.Origin

Dojo.Rotational

Dojo.Shape

Dojo.Translational

Mechanism

Missing docstring.

Missing docstring for Mechanism. Check Documenter's build log for details.

Bodies

Missing docstring.

Missing docstring for Node. Check Documenter's build log for details.

Dojo.Body — Type

Body{T} <: Node{T}

A rigid body object

id - a unique identifying number 
name - a unique identifying name 
mass - inertial property (kilograms)
inertia - inertia matrix (kilogram * meter^2)
state - representation of the system's: position, linear velocity, orientation, and angular velocity 
shape - geometry information about the body

Dojo.Origin — Type

Origin{T} <: Node{T}

Global reference frame

id - always 0 
name - always :origin 
state - defaults to zero values
shape - empty

Dojo.Shape — Type

Shape{T} 

Subtypes contain geometric and visual information for a Body{T} object.

Joints

Dojo.JointConstraint — Type

JointConstraint{T}

Joint constraint restricting translational and rotational degrees of freedom between two Body{T} objects.

id - a unique identifying number
name - a unique identifying name
translational - Translational
rotational - Rotational
spring - flag for joint springs on
damper - flag for joint dampers on
parent_id - identifying number for parent Body{T}
child_id - identifying number for child Body{T}
minimal_index - indices for minimal coordinates
impulses - joint impulses that maintain constraint between two Body{T} objects

Missing docstring.

Missing docstring for Joint. Check Documenter's build log for details.

Dojo.Rotational — Type

Rotational{T} 

Rotational component of JointConstraint 

axis - rotation axis in parent offset frame
axis_mask1 - rotation axis mask in pbody's frame
axis_mask2 - rotation axis mask in pbody's frame
axis_mask3 - rotation axis mask in pbody's frame
axis_offset - rotation axis offset from pbody's frame
spring - coefficient for joint spring
damper - coefficient for joint damper
spring_offset - nominal joint configuration
joint_limits - lower and upper limits on the joint configuration
spring_type - can be :linear or :sinusoidal (currently not implemented), if linear then we need joint_limits to avoid the 360° singularity.
input - external impulse torque

Dojo.Translational — Type

Translational{T}

Translational component of JointConstraint 

axis - translational axis in parent offset frame
axis_mask1 - translational axis mask in pbody's frame
axis_mask2 - translational axis mask in pbody's frame
axis_mask3 - translational axis mask in pbody's frame
vertices - points in parent can child frames
spring - coefficient for joint spring
damper - coefficient for joint damper
spring_offset - nominal joint configuration
joint_limits - lower and upper limits on the joint configuration
spring_type - can be :linear or :sinusoidal (currently not implemented), if linear then we need joint_limits to avoid the 360° singularity.
input - external impulse force

Contacts

Dojo.ContactConstraint — Type

ContactConstraint{T} 

Constraint object containing information for contact node.

id - unique identifying number 
name - unique identifying name 
model - type of contact model: ImpactContact, LinearContact, NonlinearContact 
parent_id - identifying number of Body experiencing contact 
child_id - always 0
impulses - contact impulses applied to Body 
impulses_dual - dual contact impulses, used by solver to enforce correct contact behaviors

Dojo.Contact — Type

Contact{T,N} 

Subtypes contain contact information for a parent Body{T} object.

Dojo.ImpactContact — Type

ImpactContact{T,N} 

Contact object for impact (i.e., no friction)

surface_normal_projector - mapping from world frame to surface tangent frame 
contact_point - position of contact on Body relative to center of mass 
offset - position of contact relative to contact_point

Dojo.LinearContact — Type

LinearContact{T,N} 

Contact object for impact and friction with a linearized friction cone

friction_coefficient - value of friction coefficient
surface_projector - mapping from world frame to surface tangent frame 
surface_normal_projector - inverse/complement of surface_projector
contact_point - position of contact on Body relative to center of mass 
offset - position of contact relative to contact_point

Dojo.NonlinearContact — Type

NonlinearContact{T,N} 

Contact object for impact and friction with a nonlinear friction cone

friction_coefficient - value of friction coefficient
surface_projector - mapping from world frame to surface tangent frame 
surface_normal_projector - inverse/complement of surface_projector
contact_point - position of contact on Body relative to center of mass 
offset - position of contact relative to contact_point

Representation

Environments

Simulation

Gradients

Visualization